+++ src/LiteObjt/liteobjt.h
@@ -220,8 +220,8 @@ extern "C" {
  */
     
  
-    void light_new__(int* PLight);
-    void light_new_go__(int* PLight, double* curtime, double* julianday, double* latitude, double* cloudcover, double* seaalbedo, double* light);
+    void light_new__(long* PLight);
+    void light_new_go__(long* PLight, double* curtime, double* julianday, double* latitude, double* cloudcover, double* seaalbedo, double* light);
   /* void light_new_getvalues__(int* plight, int* ic, int* ir, int* iz,
                 float* totallight, float* parlight,
                 float* parhorizontallight, float* hoursofsun,
